UI fixes (#45)
* Delay progression to next problem. Closes #38

* Empty lessons no longer run: closes #39

Toast also moved to drop down. It is still allowed to click on the lesson summary, but a message is displayed. This makes it clearer. Bring the toast up from the bottom can cause scrolling issues, so it now drops down from the top (not really toast anymore)

* Lesson parsing less flexible: closes #40

An introduction item now creates a new problem if either a question or intro has already been created.

* Handle images.

* Slideshow improvements

Autopause slideshow if it has a link: closes #43
Improve slide positioning: closes #44

* Completion of ui fixes.

* 1.0.26

* 1.0.26